% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{nyt_articles}
\alias{nyt_articles}
\title{Search for nytimes articles}
\usage{
nyt_articles(q = "", section = "us", page = 1, sort = "newest")
}
\arguments{
\item{q}{Search query}

\item{section}{Name of section}

\item{page}{Page number.}

\item{sort}{Sort method defaults to newest.}
}
\value{
A tibble with meta data attribute
}
\description{
Search for nytimes articles
}
