% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{nyt_comments}
\alias{nyt_comments}
\title{Get reader comments}
\usage{
nyt_comments(url, cmd = "GetCommentsAll", sort = "newest", offset = 0)
}
\arguments{
\item{url}{URL of news story}

\item{cmd}{Type of comments to return, defaults to all.}

\item{sort}{Sort method}
}
\value{
Data
}
\description{
Get reader comments
}
