% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sections.R
\docType{data}
\name{nyt_sections}
\alias{nyt_sections}
\title{Sections of the New York Times}
\format{A tibble with 30 rows and 1 variable
\describe{
  \item{section}{name of nytimes section}
}}
\source{
\url{https://www.nytimes.com/}
}
\usage{
nyt_sections
}
\description{
A dataset containing the valid sections of the nytimes.com website.
}
\keyword{datasets}
